Добавить изменения в соответствии с замечаниями. Переделать логику хранения и получения Useful.

+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -14,5 +14,5 @@ public class Review {
     private boolean isPositive;
     private long filmId;
     private long userId;
-    private Integer useful = 0;
+    private long useful = 0;
 }
